refactor: extract utils from etna.analysis.forecast.plots
